+++ src/evkx.models/Models/Performance.cs
@@ -7,21 +7,21 @@ namespace evdb.Models
 {
     public class Performance
     {
-        public int? PowerKw { get; set; }
+        public decimal? PowerKw { get; set; }
 
-        public int? PowerKwBoost { get; set; }
+        public decimal? PowerKwBoost { get; set; }
 
-        public int? TorqueNm { get; set; }  
+        public decimal? TorqueNm { get; set; }  
 
-        public int? TorqueNmBoost { get; set; }
+        public decimal? TorqueNmBoost { get; set; }
 
-        public int? BoostLengthSeconds { get; set; }
+        public decimal? BoostLengthSeconds { get; set; }
 
-        public int? TopSpeed { get; set; }
+        public decimal? TopSpeed { get; set; }
 
-        public double? ZeroToHundredKph { get; set; }
+        public decimal? ZeroToHundredKph { get; set; }
 
-        public double? ZeroToHundredKphBoost { get; set; }
+        public decimal? ZeroToHundredKphBoost { get; set; }
 
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public PerformanceOptionType? OptionType { get; set; }
